Keep track of whether the blocks and dependency list for a bug has been updated so that we only attempt to update those properties when Bug.Update is called if something has changed.
Have Bug.Update take care of updating the work time properties on a bug, this needs a couple of new
properties adding to expose the remaining time and the work time so they can be set.

Since attempting to set the work/remaining time when the logged in user is not part of the time
tracking group returns a bug edit access denied error, make the accompanying exception accept
the fault string as the message which will provide more info.